// apps/web/src/lib/api-schema.d.ts
export interface components {
    schemas: {
// ...
        ClarifyResult: {
            /** Questions */
            questions: components["schemas"]["ClarifyQuestion"][];
            /**
             * Photos
             * @default []
             */
            photos: components["schemas"]["PhotoSeen"][];
            /**
             * Voices
             * @default []
             */
            voices: components["schemas"]["VoiceHeard"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}